First time writing unit tests, we'll see...
This is the first time I've written unit tests and want to have these
available before we publish out to PyPi. So after multiple days of
wading through the question of whether or not Python allows for
importing packages or modules or whatever the word is through relative
path directories, I just decided it was better to add it directly to the
PYTHONPATH, since I'm hoping to have this pip installable before too
long.