Added some asserts
Added some asserts to the CenterOfGravity.fit() method to make sure that the options dictionary was providing options that actually existed in the columns of the frame that is passed. I have no idea if it works but I'm commiting it now so I don't have to rewrite it later.